Refactored output_provider -> logger, and moved actual output providers to io/
