+++ include/criterion/logging.h
@@ -103,7 +103,7 @@ CR_API void criterion_log(enum criterion_logging_level level, const char *msg, .
 
 # define criterion_perror(...) criterion_plog(CRITERION_IMPORTANT, CRITERION_PREFIX_ERR, __VA_ARGS__)
 
-struct criterion_output_provider {
+struct criterion_logger {
     void (*log_pre_all      )(struct criterion_test_set *set);
     void (*log_pre_suite    )(struct criterion_suite_set *set);
     void (*log_pre_init     )(struct criterion_test *test);
@@ -121,7 +121,7 @@ struct criterion_output_provider {
     void (*log_post_all     )(struct criterion_global_stats *stats);
 };
 
-extern struct criterion_output_provider normal_logging;
+extern struct criterion_logger normal_logging;
 
 CR_END_C_API
