+++ src/core/report.h
@@ -46,7 +46,7 @@ DECL_CALL_REPORT_HOOKS(POST_SUITE);
 DECL_CALL_REPORT_HOOKS(POST_ALL);
 
 #define log(Type, ...) \
-    log_(criterion_options.output_provider->log_ ## Type, __VA_ARGS__);
+    log_(criterion_options.logger->log_ ## Type, __VA_ARGS__);
 #define log_(Log, ...) \
     (Log ? Log(__VA_ARGS__) : nothing());
