+++ src/io/tap.c
@@ -26,9 +26,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include "criterion/stats.h"
-#include "criterion/logging.h"
 #include "criterion/options.h"
-#include "criterion/ordered-set.h"
 #include "compat/posix.h"
 #include "compat/strtok.h"
 #include "compat/time.h"
